+++ examples/measurement/index.html
@@ -269,8 +269,8 @@
         ],
 
         "Measuring Distances with Touch": [
-            "# Creating distance measurements with long-touch",
-            ["distance_createWithLongTouch_snapping", "Long-touch on model to create distance measurements"]
+            "# Creating distance measurements with touch input",
+            ["distance_createWithTouch_snapping", "Long-touch on model to create distance measurements"]
         ],
 
         "Measuring Distances in JavaScript": [
@@ -290,9 +290,9 @@
             ["angle_createWithMouse_nosnapping", "Click with mouse on model to create angle measurements, snapping disabled"]
         ],
 
-        "Measuring Angles with Touch [BETA]": [
-            "# Creating distance measurements with touch",
-            ["angle_createWithLongTouch_snapping", "Long-touch on model to create distance measurements"]
+        "Measuring Angles with Touch": [
+            "# Creating distance measurements with touch input",
+            ["angle_createWithTouch_snapping", "Long-touch on model to create distance measurements"]
         ],
 
         "Measuring Angles in JavaScript": [
