+++ src/plugins/AngleMeasurementsPlugin/index.js
@@ -1,4 +1,4 @@
 export * from "./AngleMeasurementsPlugin.js";
 export * from "./AngleMeasurementsControl.js";
 export * from "./AngleMeasurementsMouseControl.js";
-export * from "./AngleMeasurementsLongTouchControl.js";
+export * from "./AngleMeasurementsTouchControl.js";
