+++ src/plugins/DistanceMeasurementsPlugin/index.js
@@ -1,4 +1,4 @@
 export * from "./DistanceMeasurementsPlugin.js";
 export * from "./DistanceMeasurementsControl.js";
 export * from "./DistanceMeasurementsMouseControl.js";
-export * from "./DistanceMeasurementsLongTouchControl.js";
+export * from "./DistanceMeasurementsTouchControl.js";
